Retrieve inventory height and width for item bases
Use them in ItemBase and remove somewhat hardcoded height and width
selection.

Clean up ItemBase.

+++ WPFSKillTree/Views/CraftWindow.xaml.cs
@@ -316,12 +316,12 @@ namespace POESKillTree.Views
                 Item.ImplicitMods = msImplicitMods.GetExactMods().ToList();
             }
 
-            var quality = 0.0;
+            var quality = 0;
             if (MsQuality.Affixes != null && MsQuality.SelectedValues[0] > 0)
             {
-                quality = MsQuality.SelectedValues[0];
+                quality = (int) MsQuality.SelectedValues[0];
             }
-            Item.Properties = new ObservableCollection<ItemMod>(Item.BaseType.GetRawProperties((float) quality));
+            Item.Properties = new ObservableCollection<ItemMod>(Item.BaseType.GetRawProperties(quality));
             ApplyLocals();
 
             if (Item.IsWeapon)
